Initial file contents
We're following the structure and layout provided by http://lilypond.org/doc/v2.21/Documentation/usage/make-and-makefiles
Since this doesn't provide content for the various files (except for the makefile), we borrow that from http://lilypond.org/doc/v2.21/Documentation/snippets/staff-notation#staff-notation-orchestra-choir-and-piano-template.
Mostly we simply drop the parts present in the second but not in the first.  The one deviation from this is the oboe (present in the first but not the second) and the clarinet (vis versa).
We also add a \midi block to the all the score and part files.