Link targets and directories
The same recipe produces both the pdf and midi targets, so these need to be specified as linked targets (&:).
Also, if PDF/ and MIDI/ don't exist (as would be normal for a fresh clone of the repository), then we need to create them.  They are specified as order only prerequisites because we only want to create them if they don't exist, not condition the rebuild of pdfs based on their timestamp.